requirements.txt
streamlit==1.31.1
python-dotenv==1.0.0
replicate==0.22.0
Pillow==10.2.0
requests==2.31.0 